+++ caladrius/utils.py
@@ -9,6 +9,9 @@ import json
 
 import torch
 
+NEURAL_MODELS = ["inception", "light", "probability"]
+STATISTICAL_MODELS = ["average", "random"]
+
 # logging
 
 logging.getLogger("Fiona").setLevel(logging.ERROR)
@@ -101,7 +104,7 @@ def configuration():
     parser.add_argument(
         "--log-step",
         type=int,
-        default=100,
+        default=10,
         help="batch step size for logging information",
     )
     parser.add_argument(
@@ -113,8 +116,8 @@ def configuration():
     parser.add_argument(
         "--model-type",
         type=str,
-        default="siamese",
-        choices=["siamese", "random", "average"],
+        default=NEURAL_MODELS[0],
+        choices=NEURAL_MODELS + STATISTICAL_MODELS,
         help="type of model",
     )
 
@@ -164,23 +167,26 @@ def configuration():
         help="limit the total number of data points used, for debugging on GPU-less laptops",
     )
     parser.add_argument(
-        "--train-accuracy-threshold",
-        type=float,
-        default=0.1,
-        help="window size to calculate regression accuracy",
-    )
-    parser.add_argument(
-        "--test-accuracy-threshold",
-        type=float,
-        default=0.3,
-        help="window size to calculate regression accuracy",
-    )
-    parser.add_argument(
         "--output-type",
         type=str,
         default="regression",
         choices=["regression", "classification"],
         help="choose if want regression or classification model",
+    )
+    parser.add_argument(
+        "--selection-metric",
+        type=str,
+        default="f1_macro",
+        choices=[
+            "recall_micro",
+            "recall_macro",
+            "precision_macro",
+            "f1_macro",
+            "recall_weighted",
+            "precision_weighted",
+            "f1_weighted",
+        ],
+        help="choose metric to use for tracking best model",
     )
 
     args = parser.parse_args()
@@ -219,6 +225,8 @@ def configuration():
     arg_vars["run_report_path"] = os.path.join(
         arg_vars["checkpoint_path"], "run_report.json"
     )
+    arg_vars["statistical_model"] = arg_vars["model_type"] in STATISTICAL_MODELS
+    arg_vars["neural_model"] = arg_vars["model_type"] in NEURAL_MODELS
 
     if torch.cuda.is_available() and not arg_vars["disable_cuda"]:
         arg_vars["device"] = torch.device("cuda:{}".format(arg_vars["cuda_device"]))